+++ tests/Feature/API/CampaignsControllerTest.php
@@ -15,15 +15,15 @@ class CampaignsControllerTest extends TestCase
         WithFaker;
 
     /** @test */
-    public function a_list_of_a_workspaces_campaigns_can_be_retreived()
+    public function a_list_of_a_workspaces_campaigns_can_be_retrieved()
     {
-        $user = $this->createUserWithWorkspace();
+        [$workspace, $emailService] = $this->createUserWithWorkspaceAndEmailService();
 
-        $campaign = $this->createCampaign($user);
+        $campaign = $this->createCampaign($workspace, $emailService);
 
         $route = route('sendportal.api.campaigns.index', [
-            'workspaceId' => $user->currentWorkspace()->id,
-            'api_token' => $user->api_token,
+            'workspaceId' => $workspace->id,
+            'api_token' => $workspace->owner->api_token,
         ]);
 
         $response = $this->get($route);
@@ -40,16 +40,16 @@ class CampaignsControllerTest extends TestCase
     }
 
     /** @test */
-    public function a_single_campaign_can_be_retreived()
+    public function a_single_campaign_can_be_retrieved()
     {
-        $user = $this->createUserWithWorkspace();
+        [$workspace, $emailService] = $this->createUserWithWorkspaceAndEmailService();
 
-        $campaign = $this->createCampaign($user);
+        $campaign = $this->createCampaign($workspace, $emailService);
 
         $route = route('sendportal.api.campaigns.show', [
-            'workspaceId' => $user->currentWorkspace()->id,
+            'workspaceId' => $workspace->id,
             'campaign' => $campaign->id,
-            'api_token' => $user->api_token,
+            'api_token' => $workspace->owner->api_token,
         ]);
 
         $response = $this->get($route);
@@ -66,15 +66,20 @@ class CampaignsControllerTest extends TestCase
     /** @test */
     public function a_new_campaign_can_be_added()
     {
-        $user = $this->createUserWithWorkspace();
+        [$workspace, $emailService] = $this->createUserWithWorkspaceAndEmailService();
 
-        $route = route('sendportal.api.campaigns.store', $user->currentWorkspace()->id);
+        $route = route('sendportal.api.campaigns.store', $workspace->id);
 
         $request = [
             'name' => $this->faker->colorName,
+            'subject' => $this->faker->word,
+            'from_name' => $this->faker->word,
+            'from_email' => $this->faker->safeEmail,
+            'email_service_id' => $emailService->id,
+            'content' => $this->faker->sentence,
         ];
 
-        $response = $this->post($route, array_merge($request, ['api_token' => $user->api_token]));
+        $response = $this->post($route, array_merge($request, ['api_token' => $workspace->owner->api_token]));
 
         $response->assertStatus(201);
         $this->assertDatabaseHas('campaigns', $request);
