+++ tests/SendportalTestSupportTrait.php
@@ -4,6 +4,7 @@ namespace Tests;
 
 use Illuminate\Testing\TestResponse;
 use Sendportal\Base\Models\Campaign;
+use Sendportal\Base\Models\EmailService;
 use Sendportal\Base\Models\Segment;
 use Sendportal\Base\Models\Subscriber;
 use Sendportal\Base\Models\User;
@@ -37,10 +38,24 @@ trait SendportalTestSupportTrait {
         auth()->login($user);
     }
 
-    protected function createCampaign(User $user): Campaign
+    protected function createUserWithWorkspaceAndEmailService(): array
     {
-        return factory(Campaign::class)->create([
-            'workspace_id' => $user->currentWorkspace()->id,
+        $user = factory(User::class)->create();
+        $workspace = factory(Workspace::class)->create([
+            'owner_id' => $user->id,
+        ]);
+        $emailService = factory(EmailService::class)->create([
+            'workspace_id' => $workspace->id,
+        ]);
+
+        return [$workspace, $emailService];
+    }
+
+    protected function createCampaign(Workspace $workspace, EmailService $emailService): Campaign
+    {
+        return factory(Campaign::class)->states(['withContent', 'sent'])->create([
+            'workspace_id' => $workspace->id,
+            'email_service_id' => $emailService->id,
         ]);
     }
